# backend/src/main/java/cloudflight/integra/backend/model/Coordinates.java
package cloudflight.integra.backend.model;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.persistence.Column;
import jakarta.persistence.Embeddable;
import jakarta.validation.constraints.DecimalMax;
import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
@Embeddable
@Schema(name = "Coordinates", description = "Geographic coordinates in decimal degrees")
public class Coordinates {
    @NotNull(message = "Latitude is required")
    @DecimalMin(value = "-90.0", message = "Latitude must be between -90 and 90")
    @DecimalMax(value = "90.0", message = "Latitude must be between -90 and 90")
    @Schema(description = "Latitude in decimal degrees", example = "48.2082", minimum = "-90", maximum = "90")
    @Column(nullable = false)
    private Double latitude;

    @NotNull(message = "Longitude is required")
    @DecimalMin(value = "-180.0", message = "Longitude must be between -180 and 180")
    @DecimalMax(value = "180.0", message = "Longitude must be between -180 and 180")
    @Schema(description = "Longitude in decimal degrees", example = "16.3738", minimum = "-180", maximum = "180")
    @Column(nullable = false)
    private Double longitude;
}