# frontend/openapi.yaml.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"paths":{"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserViewDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserViewDto"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/users/unfollow/{id}":{"put":{"tags":["user-controller"],"operationId":"unfollowUser","parameters":[{"name":"followerId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserViewDto"}}}}}}},"/users/follow/{id}":{"put":{"tags":["user-controller"],"operationId":"followUser","parameters":[{"name":"followerId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserViewDto"}}}}}}},"/reviews/{id}":{"put":{"tags":["review-controller"],"operationId":"updateReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}}}},"delete":{"tags":["review-controller"],"operationId":"deleteReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/quests/{id}":{"get":{"tags":["quest-controller"],"operationId":"getQuestById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestDto"}}}}}},"put":{"tags":["quest-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestDto"}}}}}},"delete":{"tags":["quest-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/pois/{id}":{"get":{"tags":["point-of-interest-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointOfInterestResponseDto"}}}}}},"put":{"tags":["point-of-interest-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointOfInterestRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointOfInterestResponseDto"}}}}}},"delete":{"tags":["point-of-interest-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}},"/cities/{id}":{"put":{"tags":["city-controller"],"operationId":"updateCity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCityDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityDto"}}}}}},"delete":{"tags":["city-controller"],"operationId":"deleteCity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/users/auth/register":{"post":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserViewDto"}}}}}}},"/users/auth/login":{"post":{"tags":["user-controller"],"operationId":"loginUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/reviews":{"get":{"tags":["review-controller"],"operationId":"getReviews","parameters":[{"name":"poiId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"userId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewDto"}}}}}}},"post":{"tags":["review-controller"],"operationId":"createReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}}}}},"/quests":{"get":{"tags":["quest-controller"],"operationId":"getAllQuests","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestDto"}}}}}}},"post":{"tags":["quest-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestDto"}}}}}}},"/pois":{"get":{"tags":["point-of-interest-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PointOfInterestResponseDto"}}}}}}},"post":{"tags":["point-of-interest-controller"],"operationId":"savePointOfInterest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointOfInterestRequestDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointOfInterestResponseDto"}}}}}}},"/cities":{"get":{"tags":["city-controller"],"operationId":"getAllCities","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityDto"}}}}}}},"post":{"tags":["city-controller"],"operationId":"createCity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCityDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityDto"}}}}}}},"/users/{id}/following":{"get":{"tags":["user-controller"],"operationId":"getFollowing","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/UserViewDto"}}}}}}}},"/users/{id}/followers":{"get":{"tags":["user-controller"],"operationId":"getFollowers","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/UserViewDto"}}}}}}}},"/users/test-error":{"get":{"tags":["user-controller"],"operationId":"testError","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/cities/{name}":{"get":{"tags":["city-controller"],"operationId":"getCitiesByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityDto"}}}}}}}},"/cities/{name}/pois":{"get":{"tags":["city-controller"],"operationId":"getPointsOfInterestFromCityWithName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}},{"name":"poiDescription","in":"query","required":false,"schema":{"type":"string"}},{"name":"poiName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PointOfInterestResponseDto"}}}}}}}}},"components":{"schemas":{"UserBio":{"type":"object","properties":{"header":{"type":"string"},"body":{"type":"string"},"footer":{"type":"string"}}},"UserUpdateDto":{"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"bio":{"$ref":"#/components/schemas/UserBio"},"avatarUrl":{"type":"string"}}},"UserViewDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"username":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"bio":{"$ref":"#/components/schemas/UserBio"},"avatarUrl":{"type":"string"}}},"ReviewUpdateDto":{"required":["text"],"type":"object","properties":{"text":{"maxLength":1000,"minLength":0,"type":"string"},"rating":{"maximum":5,"minimum":1,"type":"integer","format":"int32"}}},"ReviewDto":{"required":["id","poiId","postedDate","text","userId"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"text":{"maxLength":1000,"minLength":0,"type":"string"},"rating":{"maximum":5,"minimum":1,"type":"integer","format":"int32"},"postedDate":{"type":"string","format":"date-time"},"userId":{"type":"string","format":"uuid"},"poiId":{"type":"string","format":"uuid"}}},"CreateQuestDto":{"required":["cityId","description","difficulty","exp","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"difficulty":{"type":"string","enum":["EASY","MEDIUM","HARD"]},"exp":{"type":"integer","format":"int32"},"cityId":{"type":"string","format":"uuid"}}},"QuestDto":{"required":["cityId","description","difficulty","exp","id","title"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"description":{"type":"string"},"difficulty":{"type":"string","enum":["EASY","MEDIUM","HARD"]},"exp":{"type":"integer","format":"int32"},"cityId":{"type":"string","format":"uuid"}}},"Coordinates":{"required":["latitude","longitude"],"type":"object","properties":{"latitude":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","description":"Latitude in decimal degrees","format":"double","example":48.2082},"longitude":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","description":"Longitude in decimal degrees","format":"double","example":16.3738}},"description":"Geographic coordinates in decimal degrees"},"PointOfInterestRequestDto":{"required":["address","cityId","coordinates","description","name","type"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"address":{"type":"string"},"type":{"type":"string","enum":["MUSEUM","PARK","CAFE"]},"cityId":{"type":"string","format":"uuid"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"PointOfInterestResponseDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"address":{"type":"string"},"type":{"type":"string","enum":["MUSEUM","PARK","CAFE"]},"cityId":{"type":"string","format":"uuid"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"UpdateCityDto":{"type":"object","properties":{"description":{"type":"string"},"imageUrl":{"type":"string"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"CityDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"country":{"type":"string"},"description":{"type":"string"},"population":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"status":{"type":"string","enum":["PENDING","APPROVED"]},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"UserRegisterDto":{"required":["email","password","username"],"type":"object","properties":{"email":{"type":"string"},"username":{"maxLength":40,"minLength":5,"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"bio":{"$ref":"#/components/schemas/UserBio"}}},"UserLoginDto":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":40,"minLength":5,"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"ReviewCreateDto":{"required":["poiId","rating","text","userId"],"type":"object","properties":{"text":{"maxLength":1000,"minLength":0,"type":"string"},"rating":{"maximum":5,"minimum":1,"type":"integer","format":"int32"},"userId":{"type":"string","format":"uuid"},"poiId":{"type":"string","format":"uuid"}}},"CreateCityDto":{"required":["coordinates","country","description","imageUrl","name","population"],"type":"object","properties":{"name":{"type":"string"},"country":{"type":"string"},"description":{"type":"string"},"population":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["PENDING","APPROVED"]},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}}}}}